sm: Fixed compilation warnings and fixed a bug when loading a saved game.

+++ semestralka/src/CardPrintStorage.h
@@ -4,7 +4,7 @@
 class CardPrintStorage{
 public:
     CardPrintStorage(int max):size(0), maxStorageSize(max){};
-    shared_ptr<Card> viewCard(int cardNO) const; //To see which card the player has chosen to play, but not actually play it(indexing from 1) 
+    shared_ptr<Card> viewCard(unsigned long cardNO) const; //To see which card the player has chosen to play, but not actually play it(indexing from 1) 
     void print() const;
     void printCensored() const;
     size_t getSize() const{
